Fix quoting problem in sqlite_connector source
When identifying the column names in a table, the code currently parses
the table_sql which has the column names quoted, so "id" is parsed in
such a way that the field name is eventually "\"id\"".

This PR strips the "" from the column names in the table_sql so that by
the time they are used in a schea, the field name is "id" and not
"\"id\"".

There may be easier ways to achieve this.